+++ Scripts/ParseEmailFiles/ParseEmailFiles.py
@@ -43,6 +43,7 @@ reload(sys)
 sys.setdefaultencoding('utf8')  # pylint: disable=no-member
 
 MAX_DEPTH_CONST = 3
+IS_NESTED_EML = False
 
 """
 https://github.com/vikramarsid/msg_parser
@@ -408,11 +409,11 @@ class EmailFormatter(object):
             if maintype == 'text' or "message" in maintype:
                 attach = MIMEText(data, _subtype=subtype)
             elif maintype == 'image':
-                attach = MIMEImage(data, _subtype=subtype)  # type: ignore
+                attach = MIMEImage(data, _subtype=subtype)  # type: ignore[assignment]
             elif maintype == 'audio':
-                attach = MIMEAudio(data, _subtype=subtype)  # type: ignore
+                attach = MIMEAudio(data, _subtype=subtype)  # type: ignore[assignment]
             else:
-                attach = MIMEBase(maintype, subtype)  # type: ignore
+                attach = MIMEBase(maintype, subtype)  # type: ignore[assignment]
                 attach.set_payload(data)
 
                 # Encode the payload using Base64
@@ -2933,7 +2934,7 @@ class Message(object):
         if property_value:
             property_detail = {property_name: property_value}
         else:
-            property_detail = None  # type: ignore
+            property_detail = None  # type: ignore[assignment]
 
         return property_detail
 
@@ -3159,7 +3160,7 @@ def parse_email_headers(header, raw=False):
     if raw:
         return headers
 
-    email_address_headers = {  # type: ignore
+    email_address_headers = {  # type: ignore[var-annotated]
         "To": [],
         "From": [],
         "CC": [],
@@ -3416,6 +3417,7 @@ def unfold(s):
 
 def handle_eml(file_path, b64=False, file_name=None, parse_only_headers=False, max_depth=3):
     global ENCODINGS_TYPES
+    global IS_NESTED_EML
 
     if max_depth == 0:
         return None, []
@@ -3517,6 +3519,7 @@ def handle_eml(file_path, b64=False, file_name=None, parse_only_headers=False, m
                         try:
                             f.write(file_content)
                             f.close()
+                            IS_NESTED_EML = True
                             inner_eml, inner_attached_emails = handle_eml(file_path=f.name,
                                                                           file_name=attachment_file_name,
                                                                           max_depth=max_depth - 1)
@@ -3569,10 +3572,12 @@ def handle_eml(file_path, b64=False, file_name=None, parse_only_headers=False, m
 
             elif part.get_content_type() == 'text/plain':
                 text = get_utf_string(part.get_payload(decode=True), 'TEXT')
-
         email_data = None
-        # if we are parsing a singed attachment it is a wrapper and we can ignore the outter "email"
-        if 'multipart/signed' not in eml.get_content_type():
+        # if we are parsing a signed attachment there can be one of two options:
+        # 1. it is a wrapper and we can ignore the outer "email"
+        # 2. it is not a wrapper and will not get into recursion, therefore we need the second condition
+        if 'multipart/signed' not in eml.get_content_type()\
+                or ('multipart/signed' in eml.get_content_type() and IS_NESTED_EML is False):
             email_data = {
                 'To': extract_address_eml(eml, 'to'),
                 'CC': extract_address_eml(eml, 'cc'),