+++ Integrations/AWS-IAM/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+
+## [19.10.2] - 2019-10-29
   - Added the following commands.
     - ***aws-iam-get-account-password-policy*** 
     - ***aws-iam-update-account-password-policy***