+++ Integrations/Panorama/CHANGELOG.md
@@ -1,6 +1,27 @@
 ## [Unreleased]
 
 
+## [19.12.0] - 2019-12-10
+  - Fixed an issue where the status log queries that returned zero results did not update to *Completed*.
+  - Added 2 commands.
+    - ***panorama-get-url-category-from-cloud***
+    - ***panorama-get-url-category-from-host***
+  - Added support to get, create, and edit custom URL category objects, including using the categories attribute in PAN-OS v9.x and above.
+
+
+## [19.11.1] - 2019-11-26
+  - Added support for a list of *job_id* in the ***panorama-query-logs*** and ***panorama-check-logs-status*** commands.
+  - Added the *ip* argument in the ***panorama-query-logs*** command.
+
+
+## [19.11.0] - 2019-11-12
+  - Fixed an issue where the ***panorama-custom-block-rule*** failed when trying to block an EDL or an address group object.
+  - Changed the *url* argument from equals to contains in the ***panorama-log-query*** command.
+  - Improved descriptions in the ***panorama-move-rule*** command.
+
+## [19.10.2] - 2019-10-29
+Added the ***panorama-security-policy-match*** command.
+
 ## [19.9.1] - 2019-09-18
 - Added the *tag* argument to several commands.
     - List commands - filter by a tag.